Breaking up package-level SpecTest into tests on its respective classes.
